Refactor FlashcardController with custom hook
